Retrieve the configuration based on the updated credential values
Since the program was retrieving the configurations by using the saved public key and private key pair, and it is not reading the modified the data before the user hit the "save" button.
Therefore used Tool class to get the modified and unsaved data before read from the configuration table.